IS-2239: Edit pagetitle on nokkelinformasjon page
